More scripture ref comparer to separate file; sort rows
